# swagger.json
{"info":{"title":"Platform Challenge","version":"1.0.0"},"host":"localhost:8080","basePath":"/","swagger":"2.0","paths":{"/":{"get":{"description":"Gets package information","produces":["application/json"],"responses":{"200":{"description":"returns object with basic info"}}}},"/events":{"get":{"description":"Returns all events","produces":["application/json"],"responses":{"200":{"description":"returns event object"}}},"post":{"description":"Posts new events","produces":["application/json"],"parameters":[{"name":"event","in":"body","description":"adds new event","required":true,"schema":{"$ref":"#/definitions/NewEvent"}}],"responses":{"201":{"description":"returns event object"}}}},"/events/{event_id}":{"get":{"description":"Returns single event by ID","produces":["application/json"],"parameters":[{"name":"event_id","in":"path","type":"string","required":true}],"responses":{"200":{"description":"returns event object"},"404":{"description":"Error with invalid event_id"}}},"put":{"description":"Updates events by ID","produces":["application/json"],"parameters":[{"name":"event_id","in":"path","type":"string","required":true},{"name":"event","in":"body","description":"updated event","required":true,"schema":{"$ref":"#/definitions/UpdatedEvent"}}],"responses":{"200":{"description":"returns updated event object"}}},"delete":{"description":"Deletes event by ID","produces":["application/json"],"parameters":[{"name":"event_id","in":"path","type":"string","required":true}],"responses":{"202":{"description":"event accepted for deletion"}}}},"/secrets":{"get":{"description":"Returns single event by ID","produces":["application/json"],"responses":{"200":{"description":"event object"},"401":{"description":"Error invalid credentials"}}}}},"definitions":{"NewEvent":{"type":"object","properties":{"name":{"type":"string"},"start":{"type":"string"},"end":{"type":"string"}}},"UpdatedEvent":{"type":"object","properties":{"name":{"type":"string"},"start":{"type":"string"},"end":{"type":"string"}}}},"responses":{},"parameters":{},"securityDefinitions":{},"tags":[]}